start with decay

+++ lib/gokart-core/src/types.rs
@@ -1,4 +1,4 @@
-pub type Int = i32;
+pub type Int = i64;
 pub type Tag = u32;
 pub type Ref = u32;
 pub type Label = usize;
